Navbar Updated
Created a proper navbar with a dropdown login. only problem is that entering random stuff accesses the login page even though user isnt logged in and no web token
