+++ client/src/App.css
@@ -36,3 +36,8 @@
     transform: rotate(360deg);
   }
 }
+
+.dropdown-menu{
+  width: 220px !important;
+
+}
