+++ client/src/App.js
@@ -3,15 +3,16 @@ import { Routes, Route} from "react-router-dom";
 
 import Dashboard from './view/Dashboard';
 import Register from './view/Register';
-import Login from './view/Login';
+import LoggedinDash from './view/LoggedinDash';
+
 
 function App() {
   return (
     <div className="App">
         <Routes>
-            <Route  path="user/register" element={<Register/>}/>
-            <Route path="/" element={<Login/>} />
-            <Route path="user/dashboard" element={<Dashboard/>} />
+            <Route  path="users/register" element={<Register/>}/>
+            <Route path="/" element={<Dashboard/>} />
+            <Route path="/Logged" element={<LoggedinDash/>} />
             <Route  path="*" element={<h1>404- Not found!</h1>}/>         
         </Routes>
     </div>
