+++ client/src/components/User.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const User = () => {
+
+    return (
+        <div>
+            <h2>User Logged In</h2>
+        </div>
+      )
+  };
+  
+  export default User;
