+++ server/package.json
@@ -17,6 +17,7 @@
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.3.0",
+    "nodemon": "^2.0.22",
     "validator": "^13.9.0"
   }
 }
